requirements.txt
asgiref==3.5.2
cffi==1.15.0
cryptography==37.0.2
Django==4.0.5
djangorestframework==3.13.1
pycparser==2.21
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
